Output

The file src/Output.jl contains the following functions that post-process numerical results.

MembraneAleFem.generate_state — Function

generate_state(path::String, t_id::Int64)

Generate the membrane state from the given path and time id.

MembraneAleFem.generate_output — Function

generate_output(mesh::Mesh, xms::Matrix{Float64}, cps::Matrix{Float64})

Generate visualization output for the given mesh, positions, and unknowns.

MembraneAleFem.analyze_output — Function

analyze_output(path::String, t_id::Int64)

Return user-specified information, with which to analyze the chosen data.

Currently, we do not have separate functions for different scenarios, and so it is up to the user to decide how they would like to analyze their data.